[reverse] Add safety check for invoking binary.

+++ src/daemon.rs
@@ -10,10 +10,10 @@ use std::{
     env, fs,
     io::prelude::*,
     os::unix::net::UnixStream,
-    path::Path,
+    path::{Path, PathBuf},
     process::{exit, id},
 };
-use sysinfo::{System, SystemExt};
+use sysinfo::{ProcessExt, System, SystemExt};
 use tokio::select;
 use tokio::time::Duration;
 use tokio::time::{sleep, Instant};
@@ -81,10 +81,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
         let mut sys = System::new_all();
         sys.refresh_all();
-        for (pid, _) in sys.processes() {
-            if pid.to_string() == swhkd_pid {
+        for (pid, process) in sys.processes() {
+            if pid.to_string() == swhkd_pid && process.exe() == env::current_exe().unwrap() {
                 log::error!("Swhkd is already running!");
-                log::error!("Pid of existing swhkd process: {}", pid.to_string());
+                log::error!("pid of existing swhkd process: {}", pid.to_string());
                 log::error!("To close the existing swhkd process, run `sudo killall swhkd`");
                 exit(1);
             }
